HADOOP-3521. Reverted the missing cast to float for sending Counters' values to Hadoop metrics which was removed by HADOOP-544.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/core/trunk@679578 13f79535-47bb-0310-9956-ffa450edef68

+++ src/mapred/org/apache/hadoop/mapred/JobInProgress.java
@@ -233,7 +233,7 @@ class JobInProgress {
       jobMetrics.setTag("group", group.getDisplayName());
       for (Counters.Counter counter : group) {
         jobMetrics.setTag("counter", counter.getDisplayName());
-        jobMetrics.setMetric("value", counter.getCounter());
+        jobMetrics.setMetric("value", (float) counter.getCounter());
         jobMetrics.update();
       }
     }
